Delete README.md
Israel Technique for Water Conservation using smart sensor and IoT technology based drip irrigation powered by solar panel 

Objective- The overall project objective is to improve water productivity and to conserve water i.e. producing more crop per drop. The goal will be achieved through increasing delivery efficiency by providing knowledge, adopting improved irrigation practices, and effective application of non-water inputs. It will contribute to increased agricultural production, more employment opportunities in rural areas, higher incomes from farming, better living standards of the farmers, and improved environment.  The project will have following key objectives:

• Improving productivity of irrigation water by efficient conveyance and its effective farm level use by adopting conservation agricultural practices and Water Conservation System 

• Production of more profitable crops through high efficiency irrigation systems for meeting increasing domestic demand and enhancing exports

• Provide Employment in Rural Areas 

• Capacity building of stakeholders in better managing irrigation water for attaining higher crop yields with less production costs


Problem- 
India has been witnessing growing water scarcity over the past several years now as population growth, industrialisation and climate change are having an impact on the country's available water sources. Pollution of water sources, which is on the rise, is also leading to water scarcity in several parts of the country.
The scarcity of water in India is leading to many problems. We are not just dependent on farmers, they have our balls in their hands. We cannot do without eating, we need food for growth,energy etc. Just as fuel is important to vehicles so is food to our body. Agriculture is the propagation of crops and rearing of animals for human consumption. The world population is increasing day by day. By 2050,we will be 9 billion. Limited resources, same size of the earth. Advances in technology has made mechanised agriculture possible.Farmers are unable to manage their crop cycles due to unavailability of water, both surface and groundwater and many other factors such as lack of knowledge as to when to water the plants and how much water is sufficient for the crops and as a result they waste a lot of their time and effort. Several farmers have reportedly committed suicides, particularly due to several factors that include unavailability of water. 

We have come up with an idea to eradicate this problem through our Project Titled - Israel Technique for Water Conservation using smart sensor and IoT technology based drip irrigation powered by solar panel 

Function-
1.) The growth of the crop depends on the type of soil, soil quality, water supply and the nutrients, fertilizers provided to the soil for crop production.
We will be using different sensors in this project. 
A Humidity Sensor, Moisture Sensor and Temperature Sensor senses, measures and reports the relative humidity, moisture and temperature of the soil. Relative humidity is the ratio of actual moisture in the soil to the highest amount of moisture that can be held at that soil temperature. 
We will collect data from the above sensors and will further compare it with the data provided by research and according to that water and fertilizers will be supplied to the farmers. 

There are different Soil conditions- 
Dry Soil
Soil with Little water 
Soil which contains sufficient water

• Dry Soil- Dry soil is the soil which is dried. The Moisture Sensor will sense the condition of soil and in dry soil full supply of water will be provided
• Soil with Little water- In this type of soil little amount of water is already present. So Water with Fertilizers will be supplied to this condition of soil after beind sensed by the Sensor
• Soil which contains sufficient water- This type of soil is already blessed with a good amount of water and moisture content in it. Less quantity of water and more quantity of fertilizer will be supplied to this type. 

2.) GSM-
This project proposes to develop a system that makes use of GSM and SMS technologies to allow for remote access to the Water Conservation, thereby making a farmer’s work much easier and less dependent of the conditions present.
With the help of GSM the present condition of the Current, Pump and Soil will be sent to the Farmer by using SMS or Text Message Technology is both English and Hindi for their Suitability to understand and read the message. 

3.)GPS- 
With the help of GPS Monitoring System, the present location of the Pump will be tracked at the time of any problem to provide ease of work to the farmers. 

4.) Solar Panel-
We will use Solar Panel as a Power Source in this Project. 

5.) Current Transformer-
The main purpose of using Current Transformer is to protect the Pump from burning. 
Current Transformer will be connected to the pump and sense the condition of the well. When there will be no water left in the well the current conception of the pump will increase then we will read it and the power of the pump will be cut and a message will be sent to the farmers using GSM to alert them. 

Outcome Of The Project-
Here, we have used different sensors like Moisture Sensor, Temperature Sensor and Humidity sensor, GPS, GSM, Current Transformer and Solar Panel as a Power Source to create an Water Conservation Smart Sensor based System powered by solar panel.
This will automatically save and conserve water which is wasted in the agriculture fields unintentionally , provide good crop production to the farmers and save time at the same time.